feat: begin epoque diffusor

+++ com/aoc/scheduler/Scheduler.java
@@ -5,5 +5,5 @@ import java.util.concurrent.Future;
 
 public interface Scheduler {
     public Future<Void> enqueue(Callable<Void> callable);
-    public void setDelay(int delay);
+    public void setDelay(long delay);
 }
